+++ ChangeLog.md
@@ -8,6 +8,7 @@
   ([#15](https://github.com/davep/hike/pull/15))
 - Made `chdir` a lot less fussy about the path given.
   ([#18](https://github.com/davep/hike/pull/18))
+- Added `quit` as a command that the comment line understands.
 
 ## v0.2.0