+++ ChangeLog.md
@@ -9,7 +9,9 @@
 - Made `chdir` a lot less fussy about the path given.
   ([#18](https://github.com/davep/hike/pull/18))
 - Added `quit` as a command that the command line understands.
+  ([#24](https://github.com/davep/hike/pull/24))
 - Added `dir` and `ls` as aliases for `chdir`.
+  ([#24](https://github.com/davep/hike/pull/24))
 
 ## v0.2.0