+++ ChangeLog.md
@@ -12,6 +12,8 @@
   ([#24](https://github.com/davep/hike/pull/24))
 - Added `dir` and `ls` as aliases for `chdir`.
   ([#24](https://github.com/davep/hike/pull/24))
+- Added `contents` as a command that the command line understands.
+  ([#24](https://github.com/davep/hike/pull/24))
 
 ## v0.2.0