+++ ChangeLog.md
@@ -14,6 +14,10 @@
   ([#24](https://github.com/davep/hike/pull/24))
 - Added `contents` as a command that the command line understands.
   ([#24](https://github.com/davep/hike/pull/24))
+- Added `bookmarks` as a command that the command line understands.
+  ([#24](https://github.com/davep/hike/pull/24))
+- Added `history` as a command that the command line understands.
+  ([#24](https://github.com/davep/hike/pull/24))
 
 ## v0.2.0