+++ ChangeLog.md
@@ -22,6 +22,8 @@
   ([#24](https://github.com/davep/hike/pull/24))
 - Added `help` as a command that the command line understands.
   ([#24](https://github.com/davep/hike/pull/24))
+- Changed command line completion so that if history is empty, known
+  commands are suggested. ([#29](https://github.com/davep/hike/pull/29))
 
 ## v0.2.0